feat: Refactor metrics to abstraction
The commit adds a new folder named "Distance" and a new class named "AbstractDistanceCalculator" to the Neighborly project. The "AbstractDistanceCalculator" class is an abstract base class that implements the "IDistanceCalculator" interface. It provides a method to calculate the distance between two vectors. This change is necessary to support different distance metrics in the Neighborly vector database.